backend/MadisonTime/requirements.txt
asgiref==3.7.2
cffi==1.15.1
cryptography==45.0.7
Django==3.2.25
django-allauth==0.63.6
django-braces==1.16.0
django-colorfield==0.11.0
django-widget-tweaks==1.4.12
gunicorn==21.2.0
numpy==1.21.0
Pillow==9.5.0
pycparser==2.21
PyMySQL==1.1.1
pytz==2024.2
sqlparse==0.4.4
typing_extensions==4.7.1
